requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# IA
google-generativeai>=0.8.0
google-genai>=1.0.0

# Banco de dados
sqlalchemy>=2.0.0
alembic>=1.12.0

# Banco vetorial
chromadb>=0.4.0

# Configuração
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0
